Check authority

+++ src/main/java/kz/yerakh/animaltrackerservice/service/impl/AccountServiceImpl.java
@@ -5,10 +5,11 @@ import kz.yerakh.animaltrackerservice.dto.AccountResponse;
 import kz.yerakh.animaltrackerservice.dto.AccountSearchCriteria;
 import kz.yerakh.animaltrackerservice.exception.EntryAlreadyExistException;
 import kz.yerakh.animaltrackerservice.exception.EntryNotFoundException;
+import kz.yerakh.animaltrackerservice.exception.UnauthorisedException;
 import kz.yerakh.animaltrackerservice.exception.InvalidValueException;
-import kz.yerakh.animaltrackerservice.exception.ModifyAccountException;
 import kz.yerakh.animaltrackerservice.repository.AccountRepository;
 import kz.yerakh.animaltrackerservice.repository.AnimalRepository;
+import kz.yerakh.animaltrackerservice.security.AuthenticationFacade;
 import kz.yerakh.animaltrackerservice.service.AccountService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -26,9 +27,13 @@ public class AccountServiceImpl implements AccountService {
     private final AccountRepository accountRepository;
     private final AnimalRepository animalRepository;
     private final PasswordEncoder passwordEncoder;
+    private final AuthenticationFacade authenticationFacade;
 
     @Override
-    public AccountResponse addAccount(AccountRequest payload) {
+    public AccountResponse registerAccount(AccountRequest payload) {
+        if (authenticationFacade.getUser().isPresent()) {
+            throw new UnauthorisedException("Request from authenticated user");
+        }
         try {
             Integer id = accountRepository.save(payload.firstName(), payload.lastName(), payload.email(),
                     passwordEncoder.encode(payload.password()));
@@ -58,6 +63,7 @@ public class AccountServiceImpl implements AccountService {
 
     @Override
     public AccountResponse updateAccount(Integer accountId, AccountRequest payload) {
+        validateAuthority(accountId);
         checkIfAccountExist(accountId);
         try {
             accountRepository.update(accountId, payload);
@@ -76,14 +82,22 @@ public class AccountServiceImpl implements AccountService {
 
     @Override
     public void deleteAccount(Integer accountId) {
+        validateAuthority(accountId);
         checkIfAccountConnectedToAnimal(accountId);
         checkIfAccountExist(accountId);
         accountRepository.delete(accountId);
     }
 
+    private void validateAuthority(Integer accountId) {
+        var user = authenticationFacade.getUser();
+        if (user.isPresent() && !accountId.equals(user.get().userData().accountId())) {
+            throw new UnauthorisedException("Unauthorised access");
+        }
+    }
+
     private void checkIfAccountExist(Integer accountId) {
         if (accountRepository.find(accountId).isEmpty()) {
-            throw new ModifyAccountException("Account not found");
+            throw new EntryNotFoundException();
         }
     }
